Domain driven development estructure
