+++ src/modules/users/infra/http/middlewares/ensureAuthenticated.ts
@@ -1,8 +1,8 @@
 import { Request, Response, NextFunction } from 'express';
 import { verify } from 'jsonwebtoken';
 
-import AppError from '../errors/AppError';
-import authConfig from '../config/auth';
+import AppError from '@shared/errors/AppError';
+import authConfig from '@config/auth';
 
 interface TokenPayload {
   iat: number;
