+++ src/modules/users/infra/http/routes/sessions.routes.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express';
 
-import CreateSessionService from '../services/CreateSessionService';
+import CreateSessionService from '@modules/users/services/CreateSessionService';
 
 const sessionRouter = Router();
