+++ src/modules/users/infra/http/routes/users.routes.ts
@@ -1,10 +1,12 @@
 import { Router } from 'express';
 import multer from 'multer';
-import uploadConfig from '../config/upload';
 
-import CreateUserService from '../services/CreateUserService';
-import ensureAuthenticated from '../middlewares/ensureAuthenticated';
-import UpdateUserAvatarService from '../services/UpdateUserAvatarService';
+import UpdateUserAvatarService from '@modules/users/services/UpdateUserAvatarService';
+import CreateUserService from '@modules/users/services/CreateUserService';
+
+import uploadConfig from '@config/upload';
+
+import ensureAuthenticated from '@modules/users/infra/http/middlewares/ensureAuthenticated';
 
 const usersRouter = Router();
 const upload = multer(uploadConfig);
