+++ src/modules/users/services/CreateSessionService.ts
@@ -2,9 +2,9 @@ import { getRepository } from 'typeorm';
 import { compare } from 'bcryptjs';
 import { sign } from 'jsonwebtoken';
 
-import User from '../models/User';
-import AppError from '../errors/AppError';
-import authConfig from '../config/auth';
+import User from '@modules/users/infra/typeorm/entities/User';
+import AppError from '@shared/errors/AppError';
+import authConfig from '@config/auth';
 
 interface RequestDTO {
   email: string;
