+++ src/modules/users/services/CreateUserService.ts
@@ -1,8 +1,8 @@
 import { hash } from 'bcryptjs';
 import { getRepository } from 'typeorm';
 
-import User from '../models/User';
-import AppError from '../errors/AppError';
+import AppError from '@shared/errors/AppError';
+import User from '@modules/users/infra/typeorm/entities/User';
 
 interface RequestDTO {
   name: string;
