+++ src/modules/users/services/UpdateUserAvatarService.ts
@@ -2,9 +2,9 @@ import fs from 'fs';
 import path from 'path';
 import { getRepository } from 'typeorm';
 
-import User from '../models/User';
-import uploadConfig from '../config/upload';
-import AppError from '../errors/AppError';
+import User from '@modules/users/infra/typeorm/entities/User';
+import uploadConfig from '@config/upload';
+import AppError from '@shared/errors/AppError';
 
 interface RequestDTO {
   user_id: string;
