+++ src/shared/infra/http/server.ts
@@ -1,11 +1,11 @@
 import express, { Request, Response, NextFunction } from 'express';
 import 'express-async-errors';
 
-import AppError from './errors/AppError';
-import routes from './routes';
-import uploadConfig from './config/upload';
+import uploadConfig from '@config/upload';
+import AppError from '@shared/errors/AppError';
+import routes from '@shared/infra/http/routes';
 
-import './database';
+import '@shared/infra/typeorm';
 
 const app = express();
